+++ packages/zowe-explorer-api/src/vscode/ZoweVsCodeExtension.ts
@@ -57,57 +57,6 @@ export class ZoweVsCodeExtension {
     }
 
     /**
-     * Show a message within VS Code dialog, and log it to an Imperative logger
-     * @param message The message to display
-     * @param severity The level of severity for the message (see `MessageSeverity`)
-     * @param logger The IZoweLogger object for logging the message
-     *
-     * @deprecated Please use `Gui.showMessage` instead
-     */
-    public static showVsCodeMessage(message: string, severity: MessageSeverity, logger: IZoweLogger): void {
-        void Gui.showMessage(message, { severity: severity, logger: logger });
-    }
-
-    /**
-     * Opens an input box dialog within VS Code, given an options object
-     * @param inputBoxOptions The options for this input box
-     *
-     * @deprecated Use `Gui.showInputBox` instead
-     */
-    public static inputBox(inputBoxOptions: vscode.InputBoxOptions): Promise<string> {
-        return Promise.resolve(Gui.showInputBox(inputBoxOptions));
-    }
-
-    /**
-     * Helper function to standardize the way we ask the user for credentials
-     * @param options Set of options to use when prompting for credentials
-     * @returns Instance of imperative.IProfileLoaded containing information about the updated profile
-     * @deprecated
-     */
-    public static async promptCredentials(options: PromptCredentialsOptions.ComplexOptions): Promise<imperative.IProfileLoaded> {
-        const profilesCache = ZoweVsCodeExtension.profilesCache;
-        const loadProfile = options.sessionName ? await profilesCache.getLoadedProfConfig(options.sessionName.trim()) : options.profile;
-        if (loadProfile == null) {
-            return undefined;
-        }
-        const loadSession = loadProfile.profile as imperative.ISession;
-
-        const creds = await ZoweVsCodeExtension.promptUserPass({ session: loadSession, ...options });
-
-        if (creds && creds.length > 0) {
-            loadProfile.profile.user = loadSession.user = creds[0];
-            loadProfile.profile.password = loadSession.password = creds[1];
-
-            const upd = { profileName: loadProfile?.name, profileType: loadProfile.type };
-            await (await profilesCache.getProfileInfo()).updateProperty({ ...upd, property: "user", value: creds[0], setSecure: options.secure });
-            await (await profilesCache.getProfileInfo()).updateProperty({ ...upd, property: "password", value: creds[1], setSecure: options.secure });
-
-            return loadProfile;
-        }
-        return undefined;
-    }
-
-    /**
      * Helper function to standardize the way we ask the user for credentials that updates ProfilesCache
      * @param options Set of options to use when prompting for credentials
      * @returns Instance of imperative.IProfileLoaded containing information about the updated profile