+++ packages/zowe-explorer/__tests__/__unit__/trees/uss/USSInit.unit.test.ts
@@ -11,7 +11,6 @@
 
 import * as vscode from "vscode";
 import { USSActions } from "../../../../src/trees/uss/USSActions";
-import { Profiles } from "../../../../src/configuration/Profiles";
 import { IJestIt, ITestContext, processSubscriptions } from "../../../__common__/testUtils";
 import { SharedContext } from "../../../../src/trees/shared/SharedContext";
 import { USSInit } from "../../../../src/trees/uss/USSInit";
@@ -30,17 +29,9 @@ describe("Test src/uss/extension", () => {
             _: { _: "_" },
         };
         const ussFileProvider: { [key: string]: jest.Mock } = {
-            addFavorite: jest.fn(),
-            removeFavorite: jest.fn(),
             createZoweSession: jest.fn(),
             filterPrompt: jest.fn(),
-            editSession: jest.fn(),
-            deleteSession: jest.fn(),
             rename: jest.fn(),
-            saveSearch: jest.fn(),
-            removeFavProfile: jest.fn(),
-            ssoLogin: jest.fn(),
-            ssoLogout: jest.fn(),
             onDidChangeConfiguration: jest.fn(),
             onDidCloseTextDocument: jest.fn(),
             getTreeView: jest.fn().mockReturnValue({
@@ -50,14 +41,6 @@ describe("Test src/uss/extension", () => {
             openWithEncoding: jest.fn(),
         };
         const commands: IJestIt[] = [
-            {
-                name: "zowe.uss.addFavorite",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "addFavorite"), arg: [test.value] }],
-            },
-            {
-                name: "zowe.uss.removeFavorite",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "removeFavorite"), arg: [test.value] }],
-            },
             {
                 name: "zowe.uss.addSession",
                 mock: [{ spy: jest.spyOn(ussFileProvider, "createZoweSession"), arg: [ussFileProvider] }],
@@ -87,17 +70,6 @@ describe("Test src/uss/extension", () => {
             {
                 name: "zowe.uss.fullPath",
                 mock: [{ spy: jest.spyOn(ussFileProvider, "filterPrompt"), arg: [test.value] }],
-            },
-            {
-                name: "zowe.uss.editSession",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "editSession"), arg: [test.value, ussFileProvider] }],
-            },
-            {
-                name: "zowe.uss.removeSession",
-                mock: [
-                    { spy: jest.spyOn(SharedContext, "isUssSession"), arg: [test.value], ret: true },
-                    { spy: jest.spyOn(ussFileProvider, "deleteSession"), arg: [test.value, undefined] },
-                ],
             },
             {
                 name: "zowe.uss.createFile",
@@ -133,47 +105,6 @@ describe("Test src/uss/extension", () => {
                 mock: [{ spy: jest.spyOn(test.value, "openUSS"), arg: [false, false, ussFileProvider] }],
             },
             {
-                name: "zowe.uss.saveSearch",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "saveSearch"), arg: [test.value] }],
-            },
-            {
-                name: "zowe.uss.removeSavedSearch",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "removeFavorite"), arg: [test.value] }],
-            },
-            {
-                name: "zowe.uss.removeFavProfile",
-                parm: [{ label: test.value }],
-                mock: [{ spy: jest.spyOn(ussFileProvider, "removeFavProfile"), arg: [test.value, true] }],
-            },
-            {
-                name: "zowe.uss.disableValidation",
-                mock: [
-                    {
-                        spy: jest.spyOn(Profiles, "getInstance"),
-                        arg: [],
-                        ret: { disableValidation: jest.fn() },
-                    },
-                ],
-            },
-            {
-                name: "zowe.uss.enableValidation",
-                mock: [
-                    {
-                        spy: jest.spyOn(Profiles, "getInstance"),
-                        arg: [],
-                        ret: { enableValidation: jest.fn(), disableValidation: jest.fn() },
-                    },
-                ],
-            },
-            {
-                name: "zowe.uss.ssoLogin",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "ssoLogin"), arg: [test.value] }],
-            },
-            {
-                name: "zowe.uss.ssoLogout",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "ssoLogout"), arg: [test.value] }],
-            },
-            {
                 name: "zowe.uss.pasteUssFile",
                 mock: [{ spy: jest.spyOn(USSActions, "pasteUss"), arg: [ussFileProvider, test.value] }],
             },
@@ -184,10 +115,6 @@ describe("Test src/uss/extension", () => {
             {
                 name: "zowe.uss.editAttributes",
                 mock: [{ spy: jest.spyOn(USSActions, "editAttributes"), arg: [test.context, ussFileProvider, test.value] }],
-            },
-            {
-                name: "zowe.uss.openWithEncoding",
-                mock: [{ spy: jest.spyOn(ussFileProvider, "openWithEncoding"), arg: [test.value, undefined] }],
             },
             {
                 name: "onDidChangeConfiguration",